move validate to its dir; add CheckSemVer unittest
